# Benchmark iterator, avoid redundant queue, remove managers.