--- a/allennlp/data/dataset_readers/multiprocess_dataset_reader.py
+++ b/allennlp/data/dataset_readers/multiprocess_dataset_reader.py
@@ -3,7 +3,7 @@ import glob
 import logging
 
 import numpy as np
-from torch.multiprocessing import Manager, Process, Queue, log_to_stderr
+from torch.multiprocessing import Manager, Process, Queue, Value, log_to_stderr
 
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
 from allennlp.data.instance import Instance
@@ -30,19 +30,20 @@ class logger:
 def _worker(reader: DatasetReader,
             input_queue: Queue,
             output_queue: Queue,
-            index: int) -> None:
+            active_workers: Value) -> None:
     """
     A worker that pulls filenames off the input queue, uses the dataset reader
-    to read them, and places the generated instances on the output queue.
-    When there are no filenames left on the input queue, it puts its ``index``
-    on the output queue and doesn't do anything else.
+    to read them, and places the generated instances on the output queue.  When
+    there are no filenames left on the input queue, it decrements
+    active_workers to signal completion.
     """
     # Keep going until you get a file_path that's None.
     while True:
         file_path = input_queue.get()
         if file_path is None:
-            # Put my index on the queue to signify that I'm finished
-            output_queue.put(index)
+            # Decrementing is not atomic. See https://docs.python.org/2/library/multiprocessing.html#multiprocessing.Value.
+            with active_workers.get_lock():
+                active_workers.value -= 1
             break
 
         logger.info(f"reading instances from {file_path}")
@@ -116,57 +117,58 @@ class MultiprocessDatasetReader(DatasetReader):
             def __init__(self) -> None:
                 self.manager = Manager()
                 self.output_queue = self.manager.Queue(outer_self.output_queue_size)
-                self.num_workers = outer_self.num_workers
 
             def __iter__(self) -> Iterator[Instance]:
-                # pylint: disable=protected-access
-                return outer_self._instances(file_path, self.manager, self.output_queue)
+                self.start()
+
+                # Keep going as long as not all the workers have finished.
+                while self.active_workers.value > 0:
+                    # Inner loop to minimize locking on self.active_workers.
+                    while True:
+                        try:
+                            # Non-blocking to handle the empty-queue case.
+                            yield self.output_queue.get(block=False, timeout=1.0)
+                        except Queue.Empty:
+                            # The queue could be empty because the workers are
+                            # all finished or because they're busy processing.
+                            # The outer loop distinguishes between these two
+                            # cases.
+                            break
+
+                self.join()
+
+            def start(self) -> None:
+                shards = glob.glob(file_path)
+                # Ensure a consistent order before shuffling for testing.
+                shards.sort()
+                num_shards = len(shards)
+
+                # If we want multiple epochs per read, put shards in the queue multiple times.
+                self.input_queue = self.manager.Queue(num_shards * outer_self.epochs_per_read + outer_self.num_workers)
+                for _ in range(outer_self.epochs_per_read):
+                    np.random.shuffle(shards)
+                    for shard in shards:
+                        self.input_queue.put(shard)
+
+                # Then put a None per worker to signify no more files.
+                for _ in range(outer_self.num_workers):
+                    self.input_queue.put(None)
+
+
+                self.processes: List[Process] = []
+                self.active_workers = Value('i', outer_self.num_workers)
+                for worker_id in range(outer_self.num_workers):
+                    process = Process(target=_worker,
+                                      args=(outer_self.reader, self.input_queue, self.output_queue, self.active_workers))
+                    logger.info(f"starting worker {worker_id}")
+                    process.start()
+                    self.processes.append(process)
+
+            def join(self) -> None:
+                for i, process in enumerate(self.processes):
+                    process.join()
+                    # Best effort logging. It's entirely possible it finished earlier.
+                    logger.info(f"worker {i} finished")
+                self.processes.clear()
 
         return QIterable()
-
-    def _instances(self, file_path: str, manager: Manager, output_queue: Queue) -> Iterator[Instance]:
-        """
-        A generator that reads instances off the output queue and yields them up
-        until none are left (signified by all ``num_workers`` workers putting their
-        ids into the queue).
-        """
-        shards = glob.glob(file_path)
-        # Ensure a consistent order before shuffling for testing.
-        shards.sort()
-        num_shards = len(shards)
-
-        # If we want multiple epochs per read, put shards in the queue multiple times.
-        input_queue = manager.Queue(num_shards * self.epochs_per_read + self.num_workers)
-        for _ in range(self.epochs_per_read):
-            np.random.shuffle(shards)
-            for shard in shards:
-                input_queue.put(shard)
-
-        # Then put a None per worker to signify no more files.
-        for _ in range(self.num_workers):
-            input_queue.put(None)
-
-        processes: List[Process] = []
-        num_finished = 0
-
-        for worker_id in range(self.num_workers):
-            process = Process(target=_worker,
-                              args=(self.reader, input_queue, output_queue, worker_id))
-            logger.info(f"starting worker {worker_id}")
-            process.start()
-            processes.append(process)
-
-        # Keep going as long as not all the workers have finished.
-        while num_finished < self.num_workers:
-            item = output_queue.get()
-            if isinstance(item, int):
-                # Means a worker has finished, so increment the finished count.
-                num_finished += 1
-                logger.info(f"worker {item} finished ({num_finished}/{self.num_workers})")
-            else:
-                # Otherwise it's an ``Instance``, so yield it up.
-                yield item
-
-        for process in processes:
-            process.join()
-        processes.clear()