@brendan-ai2 2019-08-07
@matt-peters, any interest in reviewing this PR?

It would also be great to do a debugging session with you at some point. There's an additional bottleneck that I'm having trouble pinning down. In short, we're not swapping and the input queue is kept full, but CPU utilization is poor. Generating a batch takes about an order of magnitude more than your target of several milliseconds. My best guess at the moment is contention, but I'm not sure how to measure that in Python. Compile CPython with suitable flags for running Valgrind?