@brendan-ai2 2019-08-10
@joelgrus, could you take a look as well? Thanks!