@brendan-ai2 2019-08-23
Thanks for the review! (I'm dragging my heels on merging due to some test discrepancies between TC and my local machine. I'll ping for another pass if the fix is anything major.)